+++ cpp/tests/spatial/polygon_bbox_test.cu
@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-#include <cudf/table/table.hpp>
-#include <cudf/table/table_view.hpp>
-
 #include <cuspatial/error.hpp>
 #include <cuspatial/polygon_bounding_box.hpp>
 
-#include <tests/utilities/base_fixture.hpp>
-#include <tests/utilities/column_utilities.hpp>
-#include <tests/utilities/column_wrapper.hpp>
-#include <tests/utilities/table_utilities.hpp>
-#include <tests/utilities/type_lists.hpp>
+#include <cudf/table/table.hpp>
+#include <cudf/table/table_view.hpp>
+
+#include <cudf_test/base_fixture.hpp>
+#include <cudf_test/column_utilities.hpp>
+#include <cudf_test/column_wrapper.hpp>
+#include <cudf_test/table_utilities.hpp>
+#include <cudf_test/type_lists.hpp>
 
 template <typename T>
 struct BoundingBoxTest : public cudf::test::BaseFixture {